requirements.txt
# Core RAG
langchain==0.2.16
langchain-community==0.2.16
groq
chromadb==0.5.5
sentence-transformers==3.0.1

# Document parsing
pypdf==4.3.1
python-docx==1.1.2
python-pptx==1.0.2

# API
fastapi==0.112.0
uvicorn==0.30.6
python-multipart==0.0.9

# Storage & utilities
sqlalchemy==2.0.32
pydantic==2.8.2
python-dotenv==1.0.1
